+++ src/main/java/com/medkandirou/youwatch/category/ICategorie.java
@@ -2,5 +2,5 @@ package com.medkandirou.youwatch.category;
 
 import com.medkandirou.youwatch.helpers.IData;
 
-public interface ICategorie extends IData<CategoryDTOreq, CategoryDTOres,Long> {
+public interface ICategorie extends IData<CategoryDTOreq, CategoryDTOres,Integer> {
 }
